Working ASA and Fix for Odd Bug
Odd Bug - Because the code for ASA executes asynchronously from the normal Unity workflow, a bug can occur where the script calls upon an object that does not yet exist. For now, I solved the bug by referencing the needed objects a second time latter in the code for the SpatialAnchorManager and the PreviewMenuController scripts. (See the "Research and Development"->"Spatial Anchors" -> "Working Method" section of the ReadMe for more info.)